Removed styles management.
Now the styles will be managed solely through the already minified "src/styles.js" file.
The inclusion of the demo style file in the JavaScript demo files was replaced with  an HTML <link> inclusion in the "index.html" file.
The styles management packages and configurations were removed, but the package lock file was not updated as previously the yarn package manager was used and I'm switching to the npm package manager and I do not want yet to install packages as more packages will be soon removed during removal of all tests.