Removed the copying of the production files feature.
The "copy-webpack-plugin" was removed and its configuration in the "webpack.prod.config.js" file.
The package to the npm, Inc.'s registry will be now published from the "package" directory. So the "webpack.prod.config.js" file was configured to output the production file to "package/index.js". The "package.json" file was copied to the "package" directory. So in the project there will be two "package.json" files—one in the project's main directory, and the other in the "package" directory (the "package" directory is named as such to make the creation of an installable tarball archive easier as such archive's files must be located in the directory named so). The project's main "package.json" directory was trimmed of all data in spite of dependencies information and scripts declaration—such data will be only in the "package/package.json" file to not create a need to, for example, specify the package version in both "package.json" files; so, thanks to this, the primary change of this commit—that is, the removing of the copying feature—can be made. Furthermore, thanks to such configuration, the ".npmignore" file is unneeded and was therefore removed (the npm publish command will be run from the "package" directory). The chief drawback is that the "README.md" file must as well be in two directories and any updates to must be copied manually; the file wasn't yet copied yet as it calls for a professionalization of its English—it will be done in a separate commit after the refurbishment is done.