Added the "package-lock.json" file.
The package manager was switched to yarn, but the removal of the "yarn.lock" file is already done.
The packages weren't yet installed in my local repository so now, after the recent commits, where a lot of packages was removed, the packages were installed, and the "package-lock.json" file is published.