Removed the last "data-" prefixed attribute and removed the removing of it.
When all tests were removed the "data-" prefixed attributes used by the Cypress framework weren't removed. After the changing of the stage system the most of them were removed and only one was left. Now it was removed and the configuration of the webapack was updated and the webpack plugin used to remove it was uninstalled.